Docker: fold StageGraph into the one recipe that reads it
